TaskBasedProduction.jl

Package for task-based production model.

Package Features

Unit input demand calculator

General unit input demand calculation

General production function evaluation

Specific production function evaluation

Marginal productivity of labor calculation

Elasticity of substitution and complementarity calculation

General marginal productivity of labor calculation

Elasticity of substitution and complementarity calculation general case

Function Documentation

TaskBasedProduction.unitInputDemand — Function

unitInputDemand(xT::AbstractArray{<:Real}, θ::Real, κ::Real, z::Real, αVec::AbstractArray{<:Real}, skipParamChecks::Bool = false) -> AbstractArray{<:Real}

Calculates unit labor demands given blueprint scale θ, blueprint shape κ, productivity z, an array of comparative advantage values αVec with H elements (one for each worker type), and an array xT of H-1 thresholds in task space.

Arguments

xT: An array of H-1 thresholds in task space.

θ: Blueprint scale parameter.

κ: Blueprint shape parameter.

z: Productivity parameter.

αVec: An array of comparative advantage values with H elements.

skipParamChecks: A boolean indicating whether to skip parameter checks (default is false).

Returns

An array representing the labor demand for each labor type.

TaskBasedProduction.unitInputDemand_general — Function

unitInputDemand_general(xT::Vector{Float64}, z::Real, b_g::Function, e_h::Vector{Function}) -> Vector{Float64}

Calculates unit labor demands given an array xT of H-1 thresholds in task space, a productivity value z, a density function b_g for the task distribution, and an array e_h of H functions representing the cost of each labor type as a function of task complexity.

The function first verifies that b_g is a valid density function. Then it computes the labor demand for each labor type by numerically integrating the ratio b_g(x) / (z * e_h[h](x)) over the intervals defined by the thresholds in xT.

Arguments

xT: A vector of H-1 thresholds in task space.

z: Productivity value.

b_g: A density function for the task distribution.

e_h: A vector of H functions representing the cost of each labor type as a function of task complexity.

Returns

A vector representing the labor demand for each labor type.

TaskBasedProduction.prod_fun_general — Function

prodfungeneral(l::AbstractArray{<:Real}, z::Real, bg:: Function, eh::Vector{Function})

Calculates the quantity produced (q), and task thresholds (xT) given labor inputs (labor_input), productivity z, and general blueprint density function and a vector of efficiency functions, one for each labor type.

Inputs:

l: Array of labor inputs of different types.

z: Productivity parameter.

b_g: Blueprint density function

e_h: Vector of efficiency functions, one for each type

Returns:

q: Quantity produced.

xT: Array of task thresholds.

TaskBasedProduction.prod_fun — Function

prod_fun(l::AbstractArray{<:Real}, θ::Real, κ::Real, z::Real, αVec::AbstractArray{<:Real})

Calculates the quantity produced (q), and task thresholds (xT) given labor inputs (l), blueprint scale θ, blueprint shape κ, productivity z, and an array of comparative advantage values αVec with H elements (one for each worker type).

Inputs:

l: Array of labor inputs of different types.

θ: Blueprint scale parameter.

κ: Blueprint shape parameter.

z: Productivity parameter.

αVec: Array of comparative advantage values with H elements.

Returns:

q: Quantity produced.

xT: Array of task thresholds.

TaskBasedProduction.margProdLabor — Function

margProdLabor(labor_input::AbstractArray{<:Real}, θ::Real, κ::Real, z::Real, αVec::AbstractArray{<:Real}) -> AbstractArray{<:Real}

Calculates the marginal productivity of labor for each worker type given the input parameters.

Arguments

labor_input: An array of labor demand values.

θ: Blueprint scale parameter.

κ: Blueprint shape parameter.

z: Productivity parameter.

αVec: An array of comparative advantage values.

Returns

An array representing the marginal productivity of labor for each worker type.

TaskBasedProduction.elasticity_sub_comp — Function

elasticity_sub_comp(labor_input::AbstractArray{<:Real}, θ::Real, κ::Real, z::Real, αVec::AbstractArray{<:Real}) -> (AbstractArray{<:Real}, AbstractArray{<:Real})

Calculates the elasticity of substitution and complementarity for a given set of parameters.

Arguments

labor_input: An array of labor inputs of different types with H elements.

θ: Blueprint scale parameter.

κ: Blueprint shape parameter.

z: Productivity parameter.

αVec: An array of comparative advantage values with H elements.

Returns

ϵ_h_sub: Matrix of elasticity of substitution values for each worker type h (rows) relative to worker type h_prime (columns).

ϵ_h_compl: Matrix of elasticity of complementarity values for each worker type h (rows) relative to worker type h_prime (columns).

TaskBasedProduction.margProdLabor_general — Function

margProdLabor_general(labor_input::AbstractArray{<:Real}, z::Real, b_g::Function, e_h::Vector{Function}) -> AbstractArray{<:Real}

Calculates the marginal productivity of labor for each worker type given the input parameters.

Arguments

labor_input: An array of labor inputs of different types with H elements.

z: A productivity scalar.

b_g: A task density function.

e_h: A vector of comparative advantage functions.

Returns

An array representing the marginal productivity of labor for each worker type.

TaskBasedProduction.elasticity_sub_comp_general — Function

elasticity_sub_comp_general(labor_input::AbstractArray{<:Real}, z::Real, b_g::Function, e_h::Vector{Function}) -> (AbstractArray{<:Real}, AbstractArray{<:Real})

Calculates the elasticity of substitution and complementarity for a given set of parameters.

Arguments

labor_input: An array of labor inputs of different types with H elements.

z: Productivity parameter.

b_g: General task density function.

e_h: Vector of comparative advantage functions.

Returns

ϵ_h_sub: Matrix of elasticity of substitution values for each worker type h (rows) relative to worker type h_prime (columns).

ϵ_h_compl: Matrix of elasticity of complementarity values for each worker type h (rows) relative to worker type h_prime (columns).